Use the string_enum! macro across all of crater
